Refactor headers in panda/board/jungle/*.h
